Move the Projects.jsx to a separate folder 'projects'.
