+++ src/components/form/ui/Form.jsx
@@ -5,7 +5,7 @@ import SocialMedia from "../components/SocialMedia";
 import Summary from "../components/Summary";
 import Education from "../components/Education";
 import WorkExperience from "../components/WorkExperience";
-import Projects from "../components/Projects";
+import Projects from "../components/projects/ui/Projects";
 import Skills from "../components/skills/ui/Skills";
 import Languages from "../components/languages/ui/Languages";
 import TestsAndCertifications from "../components/testsAndCertifications/ui/TestsAndCertifications";
